Implement logging and begin move towards Unix-specific functions
Add a simple mechanism for writing lines of text to a file on disk in a
(relatively) thread-safe way. This uses the Unix file syscalls (`open`, `write`,
etc.) instead of the stdio versions. In the spirit of OS-dependency, rework the
allocation functions to use `mmap`/`munmap` instead of the stdlib equivalents as
well.