Parse/store request cookies, and clean up login validation
Read any cookies provided by the CGI metavariable HTTP_COOKIE, and store them in
the same way as url and form parameters. To parse the cookie string, update the
Key_Value_Pair handling so that the delimiter is parameterized.

Consolidate the login and registration validation into their dedicated
functions, in preparation for reading/setting the session ID cookie.