Add handling for key/value pairs without values
Previously, `consume_key_value_pair()` returned a value of 0 in cases where only
a key could be found. This caused a bug in code that expected value to be a
string (e.g. so that `string_length()` could be called on it safely).

Update the function to return an empty string value instead in such cases, until
such a time it is determined that distinguishing between null and empty string
values is worthwhile. This way the caller is not responsible for checking for
null values, only keys.